bug 875062 - move and update tests
